backend/services/donation_processing_service/pyproject.toml
[project]
name = "donation-processing-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "requests>=2.32.4",
]
